move include dir on same level as section
