+++ include/LuaCFuncs.h
@@ -1,0 +1,1231 @@
+#pragma once
+
+#include "LuaAPI.h"
+
+#define LuaC(addr, name) \
+  int FA_##name(lua_State *L) asm(#addr);
+
+LuaC(0x416490, BeginLoggingStats)
+LuaC(0x416640, EndLoggingStats)
+LuaC(0x41B790, GetCommandLineArg)
+LuaC(0x41BA20, HasCommandLineArg)
+LuaC(0x41CB60, ConExecute)
+LuaC(0x41D180, ConExecuteSave)
+LuaC(0x41EB00, AddConsoleOutputReciever)
+LuaC(0x41ED00, RemoveConsoleOutputReciever)
+LuaC(0x456BE0, DiskFindFiles)
+LuaC(0x456E30, DiskGetFileInfo)
+LuaC(0x457160, DiskToLocal)
+LuaC(0x4572E0, Basename)
+LuaC(0x457460, Dirname)
+LuaC(0x4575C0, FileCollapsePath)
+LuaC(0x4A5700, CreatePrefetchSet)
+LuaC(0x4A5810, CPrefetchSetUpdate)
+LuaC(0x4A5950, CPrefetchSetReset)
+LuaC(0x4B2EE0, EnumColorNames)
+LuaC(0x4B41E0, Trace)
+LuaC(0x4C7A90, IsDestroyed)
+LuaC(0x4C9D00, ForkThread)
+LuaC(0x4C9FE0, KillThread)
+LuaC(0x4CA890, WaitFor)
+LuaC(0x4CAA70, SuspendCurrentThread)
+LuaC(0x4CAB90, ResumeThread)
+LuaC(0x4CAD60, CurrentThread)
+LuaC(0x4CD8F0, print)
+LuaC(0x4CD9A0, LOG)
+LuaC(0x4CDA50, WARN)
+LuaC(0x4CDB00, SPEW)
+LuaC(0x4CEAF0, doscript)
+LuaC(0x4CF930, __index)
+LuaC(0x4CF9C0, __newindex)
+LuaC(0x4CFE00, Vector2)
+LuaC(0x4D0140, Vector)
+LuaC(0x4D0600, PointVector)
+LuaC(0x4D0C70, Rect)
+LuaC(0x4D0EB0, VDist3)
+LuaC(0x4D1070, VDist3Sq)
+LuaC(0x4D1230, VDist2)
+LuaC(0x4D1440, VDist2Sq)
+LuaC(0x4D1650, VDot)
+LuaC(0x4D17E0, VDiff)
+LuaC(0x4D19C0, VAdd)
+LuaC(0x4D1BA0, VMult)
+LuaC(0x4D1D70, VPerpDot)
+LuaC(0x4D1F10, MATH_IRound)
+LuaC(0x4D2030, MATH_Lerp)
+LuaC(0x4D3470, STR_xtoi)
+LuaC(0x4D3580, STR_itox)
+LuaC(0x4D3700, STR_Utf8Len)
+LuaC(0x4D3810, STR_Utf8SubString)
+LuaC(0x4D3A10, STR_GetTokens)
+LuaC(0x4D3D70, exists)
+LuaC(0x4D3EA0, GetVersion)
+LuaC(0x4D68C0, SHGetFolderPath)
+LuaC(0x4DFD90, Sound)
+LuaC(0x4DFED0, RPCSound)
+LuaC(0x4E0010, GetCueBank)
+LuaC(0x50D3B0, OrientFromDir)
+LuaC(0x50D510, EulerToQuaternion)
+LuaC(0x50D730, MinLerp)
+LuaC(0x50D950, MinSlerp)
+LuaC(0x5284D0, SpecFootprints)
+LuaC(0x528AF0, RegisterUnitBlueprint)
+LuaC(0x528BC0, RegisterPropBlueprint)
+LuaC(0x528C90, RegisterProjectileBlueprint)
+LuaC(0x528D60, RegisterMeshBlueprint)
+LuaC(0x528E20, RegisterTrailEmitterBlueprint)
+LuaC(0x528ED0, RegisterEmitterBlueprint)
+LuaC(0x528F80, RegisterBeamBlueprint)
+LuaC(0x529030, BlueprintLoaderUpdateProgress)
+LuaC(0x547030, CreateResourceDeposit)
+LuaC(0x5556B0, EntityCategory__add)
+LuaC(0x555840, EntityCategory__sub)
+LuaC(0x5559D0, EntityCategory__mul)
+LuaC(0x555B60, EntityCategoryGetUnitList)
+LuaC(0x555D70, EntityCategoryEmpty)
+LuaC(0x55B610, SecondsPerTick)
+LuaC(0x585EF0, CAiBrainIsOpponentAIRunning)
+LuaC(0x586070, CAiBrainGetArmyIndex)
+LuaC(0x5861C0, CAiBrainGetFactionIndex)
+LuaC(0x586310, CAiBrainSetCurrentPlan)
+LuaC(0x5864A0, CAiBrainGetPersonality)
+LuaC(0x5865F0, CAiBrainSetCurrentEnemy)
+LuaC(0x586770, CAiBrainGetCurrentEnemy)
+LuaC(0x5868D0, CAiBrainGetUnitBlueprint)
+LuaC(0x586AD0, CAiBrainGetArmyStat)
+LuaC(0x586DA0, CAiBrainSetArmyStat)
+LuaC(0x587020, CAiBrainAddArmyStat)
+LuaC(0x5872A0, CAiBrainSetGreaterOf)
+LuaC(0x587520, CAiBrainGetBlueprintStat)
+LuaC(0x5876E0, CAiBrainGetCurrentUnits)
+LuaC(0x587870, CAiBrainGetListOfUnits)
+LuaC(0x587B80, CAiBrainSetArmyStatsTrigger)
+LuaC(0x587FA0, CAiBrainRemoveArmyStatsTrigger)
+LuaC(0x5881F0, CAiBrainGiveResource)
+LuaC(0x5883E0, CAiBrainGiveStorage)
+LuaC(0x5885E0, CAiBrainTakeResource)
+LuaC(0x588850, CAiBrainSetResourceSharing)
+LuaC(0x5889A0, CAiBrainFindUnit)
+LuaC(0x588C10, CAiBrainFindUpgradeBP)
+LuaC(0x588EB0, CAiBrainFindUnitToUpgrade)
+LuaC(0x589380, CAiBrainDecideWhatToBuild)
+LuaC(0x589720, CAiBrainGetArmyStartPos)
+LuaC(0x589890, CAiBrainCreateUnitNearSpot)
+LuaC(0x589DB0, CAiBrainCreateResourceBuildingNearest)
+LuaC(0x58A440, CAiBrainFindPlaceToBuild)
+LuaC(0x58B3A0, CAiBrainCanBuildStructureAt)
+LuaC(0x58B610, CAiBrainBuildStructure)
+LuaC(0x58BA40, CAiBrainNumCurrentlyBuilding)
+LuaC(0x58BCB0, CAiBrainGetAvailableFactories)
+LuaC(0x58BF80, CAiBrainCanBuildPlatoon)
+LuaC(0x58C490, CAiBrainBuildPlatoon)
+LuaC(0x58C840, CAiBrainBuildUnit)
+LuaC(0x58CA40, CAiBrainIsAnyEngineerBuilding)
+LuaC(0x58CCA0, CAiBrainGetNumPlatoonsWithAI)
+LuaC(0x58CE30, CAiBrainGetNumPlatoonsTemplateNamed)
+LuaC(0x58CFC0, CAiBrainPlatoonExists)
+LuaC(0x58D140, CAiBrainGetPlatoonsList)
+LuaC(0x58D360, CAiBrainDisbandPlatoon)
+LuaC(0x58D4D0, CAiBrainDisbandPlatoonUniquelyNamed)
+LuaC(0x58D650, CAiBrainMakePlatoon)
+LuaC(0x58DC60, CAiBrainAssignUnitsToPlatoon)
+LuaC(0x58DFA0, CAiBrainGetPlatoonUniquelyNamed)
+LuaC(0x58E140, CAiBrainGetNumUnitsAroundPoint)
+LuaC(0x58E3D0, CAiBrainGetUnitsAroundPoint)
+LuaC(0x58E830, CAiBrainFindClosestArmyWithBase)
+LuaC(0x58EB40, CAiBrainSetUpAttackVectorsToArmy)
+LuaC(0x58ED60, CAiBrainGetAttackVectors)
+LuaC(0x58EF60, CAiBrainPickBestAttackVector)
+LuaC(0x58F370, CAiBrainGetEconomyStored)
+LuaC(0x58F530, CAiBrainGetEconomyStoredRatio)
+LuaC(0x58F730, CAiBrainGetEconomyIncome)
+LuaC(0x58F8F0, CAiBrainGetEconomyUsage)
+LuaC(0x58FAB0, CAiBrainGetEconomyRequested)
+LuaC(0x58FC70, CAiBrainGetEconomyTrend)
+LuaC(0x58FE30, CAiBrainGetMapWaterRatio)
+LuaC(0x58FF80, CAiBrainAssignThreatAtPosition)
+LuaC(0x590260, CAiBrainGetThreatAtPosition)
+LuaC(0x5905B0, CAiBrainGetThreatBetweenPositions)
+LuaC(0x5908D0, CAiBrainGetHighestThreatPosition)
+LuaC(0x590C00, CAiBrainGetThreatsAroundPosition)
+LuaC(0x591020, CAiBrainCheckBlockingTerrain)
+LuaC(0x5917F0, CAiBrainGetNoRushTicks)
+LuaC(0x5A5860, CAiNavigatorImplSetGoal)
+LuaC(0x5A5A90, CAiNavigatorImplSetDestUnit)
+LuaC(0x5A5BF0, CAiNavigatorImplAbortMove)
+LuaC(0x5A5D20, CAiNavigatorImplBroadcastResumeTaskEvent)
+LuaC(0x5A5E50, CAiNavigatorImplSetSpeedThroughGoal)
+LuaC(0x5A5F90, CAiNavigatorImplGetCurrentTargetPos)
+LuaC(0x5A6100, CAiNavigatorImplGetGoalPos)
+LuaC(0x5A6270, CAiNavigatorImplGetStatus)
+LuaC(0x5A63C0, CAiNavigatorImplHasGoodPath)
+LuaC(0x5A6500, CAiNavigatorImplFollowingLeader)
+LuaC(0x5A6640, CAiNavigatorImplIgnoreFormation)
+LuaC(0x5A6780, CAiNavigatorImplIsIgnorningFormation)
+LuaC(0x5A68C0, CAiNavigatorImplAtGoal)
+LuaC(0x5A6A00, CAiNavigatorImplCanPathToGoal)
+LuaC(0x5BA610, CAiPersonalityGetPersonalityName)
+LuaC(0x5BA730, CAiPersonalityGetChatPersonality)
+LuaC(0x5BA850, CAiPersonalityGetDifficulty)
+LuaC(0x5BA970, CAiPersonalityAdjustDelay)
+LuaC(0x5BAB40, CAiPersonalityGetArmySize)
+LuaC(0x5BAC80, CAiPersonalityGetPlatoonSize)
+LuaC(0x5BADC0, CAiPersonalityGetAttackFrequency)
+LuaC(0x5BAF00, CAiPersonalityGetRepeatAttackFrequency)
+LuaC(0x5BB040, CAiPersonalityGetCounterForces)
+LuaC(0x5BB180, CAiPersonalityGetIntelGathering)
+LuaC(0x5BB2C0, CAiPersonalityGetCoordinatedAttacks)
+LuaC(0x5BB400, CAiPersonalityGetExpansionDriven)
+LuaC(0x5BB540, CAiPersonalityGetTechAdvancement)
+LuaC(0x5BB680, CAiPersonalityGetUpgradesDriven)
+LuaC(0x5BB7C0, CAiPersonalityGetDefenseDriven)
+LuaC(0x5BB900, CAiPersonalityGetEconomyDriven)
+LuaC(0x5BBA40, CAiPersonalityGetFactoryTycoon)
+LuaC(0x5BBB80, CAiPersonalityGetIntelBuildingTycoon)
+LuaC(0x5BBCC0, CAiPersonalityGetSuperWeaponTendency)
+LuaC(0x5BBE00, CAiPersonalityGetFavouriteStructures)
+LuaC(0x5BBFD0, CAiPersonalityGetAirUnitsEmphasis)
+LuaC(0x5BC110, CAiPersonalityGetTankUnitsEmphasis)
+LuaC(0x5BC250, CAiPersonalityGetBotUnitsEmphasis)
+LuaC(0x5BC390, CAiPersonalityGetSeaUnitsEmphasis)
+LuaC(0x5BC4D0, CAiPersonalityGetSpecialtyForcesEmphasis)
+LuaC(0x5BC610, CAiPersonalityGetSupportUnitsEmphasis)
+LuaC(0x5BC750, CAiPersonalityGetDirectDamageEmphasis)
+LuaC(0x5BC890, CAiPersonalityGetInDirectDamageEmphasis)
+LuaC(0x5BC9D0, CAiPersonalityGetFavouriteUnits)
+LuaC(0x5BCBA0, CAiPersonalityGetSurvivalEmphasis)
+LuaC(0x5BCCE0, CAiPersonalityGetTeamSupport)
+LuaC(0x5BCE20, CAiPersonalityGetFormationUse)
+LuaC(0x5BCF60, CAiPersonalityGetTargetSpread)
+LuaC(0x5BD0A0, CAiPersonalityGetQuittingTendency)
+LuaC(0x5BD1E0, CAiPersonalityGetChatFrequency)
+LuaC(0x5C29D0, ReconBlipGetBlueprint)
+LuaC(0x5C2B30, ReconBlipGetSource)
+LuaC(0x5C2C90, ReconBlipIsSeenEver)
+LuaC(0x5C2E00, ReconBlipIsSeenNow)
+LuaC(0x5C2F70, ReconBlipIsMaybeDead)
+LuaC(0x5C30E0, ReconBlipIsOnOmni)
+LuaC(0x5C3250, ReconBlipIsOnSonar)
+LuaC(0x5C33C0, ReconBlipIsOnRadar)
+LuaC(0x5C3530, ReconBlipIsKnownFake)
+LuaC(0x5D9930, CAiAttackerImplGetUnit)
+LuaC(0x5D9A70, CAiAttackerImplAttackerWeaponsBusy)
+LuaC(0x5D9BB0, CAiAttackerImplGetWeaponCount)
+LuaC(0x5D9D00, CAiAttackerImplSetDesiredTarget)
+LuaC(0x5D9EA0, CAiAttackerImplGetDesiredTarget)
+LuaC(0x5DA000, CAiAttackerImplStop)
+LuaC(0x5DA130, CAiAttackerImplCanAttackTarget)
+LuaC(0x5DA2E0, CAiAttackerImplFindBestEnemy)
+LuaC(0x5DA490, CAiAttackerImplGetTargetWeapon)
+LuaC(0x5DA650, CAiAttackerImplGetPrimaryWeapon)
+LuaC(0x5DA7A0, CAiAttackerImplGetMaxWeaponRange)
+LuaC(0x5DA8E0, CAiAttackerImplIsWithinAttackRange)
+LuaC(0x5DACE0, CAiAttackerImplIsTooClose)
+LuaC(0x5DAE90, CAiAttackerImplIsTargetExempt)
+LuaC(0x5DB010, CAiAttackerImplHasSlavedTarget)
+LuaC(0x5DB1A0, CAiAttackerImplResetReportingState)
+LuaC(0x5DB2D0, CAiAttackerImplForceEngage)
+LuaC(0x5E3150, AITarget)
+LuaC(0x5E8640, UnitTransportDetachAllUnits)
+LuaC(0x6233C0, CUnitScriptTaskGetUnit)
+LuaC(0x6234F0, CUnitScriptTaskSetAIResult)
+LuaC(0x623660, LUnitMove)
+LuaC(0x623900, LUnitMoveNear)
+LuaC(0x631F20, CreateAimController)
+LuaC(0x632140, CAimManipulatorSetFiringArc)
+LuaC(0x632340, CAimManipulatorSetResetPoseTime)
+LuaC(0x6324B0, CAimManipulatorOnTarget)
+LuaC(0x6325F0, CAimManipulatorSetEnabled)
+LuaC(0x632730, CAimManipulatorGetHeadingPitch)
+LuaC(0x632890, CAimManipulatorSetHeadingPitch)
+LuaC(0x632A40, CAimManipulatorSetAimHeadingOffset)
+LuaC(0x634C70, CBoneEntityManipulatorSetPivot)
+LuaC(0x634E80, EntityAttachBoneToEntityBone)
+LuaC(0x636800, CreateBuilderArmController)
+LuaC(0x6369D0, CBuilderArmManipulatorSetAimingArc)
+LuaC(0x636BD0, CBuilderArmManipulatorGetHeadingPitch)
+LuaC(0x636D30, CBuilderArmManipulatorSetHeadingPitch)
+LuaC(0x637FA0, CreateCollisionDetector)
+LuaC(0x638110, CCollisionManipulatorEnableTerrainCheck)
+LuaC(0x638250, CCollisionManipulatorEnable)
+LuaC(0x638380, CCollisionManipulatorDisable)
+LuaC(0x6384C0, CCollisionManipulatorWatchBone)
+LuaC(0x639C80, CreateFootPlantController)
+LuaC(0x63BAC0, IAniManipulatorSetPrecedence)
+LuaC(0x63BC60, IAniManipulatorEnable)
+LuaC(0x63BD90, IAniManipulatorDisable)
+LuaC(0x63BEC0, IAniManipulatorDestroy)
+LuaC(0x6404B0, CreateAnimator)
+LuaC(0x640670, CAnimationManipulatorPlayAnim)
+LuaC(0x6408E0, CAnimationManipulatorGetRate)
+LuaC(0x640A20, CAnimationManipulatorSetRate)
+LuaC(0x640BA0, CAnimationManipulatorGetAnimationFraction)
+LuaC(0x640D10, CAnimationManipulatorSetAnimationFraction)
+LuaC(0x640EB0, CAnimationManipulatorGetAnimationTime)
+LuaC(0x640FF0, CAnimationManipulatorSetAnimationTime)
+LuaC(0x641160, CAnimationManipulatorGetAnimationDuration)
+LuaC(0x6415F0, CAnimationManipulatorSetBoneEnabled)
+LuaC(0x6417B0, CAnimationManipulatorSetOverwriteMode)
+LuaC(0x6418F0, CAnimationManipulatorSetDisableOnSignal)
+LuaC(0x641A30, CAnimationManipulatorSetDirectionalAnim)
+LuaC(0x643D80, CreateRotator)
+LuaC(0x644280, CRotateManipulatorSetSpinDown)
+LuaC(0x6443F0, CRotateManipulatorSetGoal)
+LuaC(0x6445C0, CRotateManipulatorClearGoal)
+LuaC(0x644710, CRotateManipulatorSetSpeed)
+LuaC(0x6448B0, CRotateManipulatorSetTargetSpeed)
+LuaC(0x644A80, CRotateManipulatorSetAccel)
+LuaC(0x644C20, CRotateManipulatorClearFollowBone)
+LuaC(0x644D90, CRotateManipulatorSetFollowBone)
+LuaC(0x644F30, CRotateManipulatorGetCurrentAngle)
+LuaC(0x6450A0, CRotateManipulatorSetCurrentAngle)
+LuaC(0x6462B0, CreateSlaver)
+LuaC(0x646490, CSlaveManipulatorSetMaxRate)
+LuaC(0x647660, CreateSlider)
+LuaC(0x647A30, CSlideManipulatorSetWorldUnits)
+LuaC(0x647B80, CSlideManipulatorSetSpeed)
+LuaC(0x647D80, CSlideManipulatorSetAcceleration)
+LuaC(0x647F20, CSlideManipulatorSetDeceleration)
+LuaC(0x6480C0, CSlideManipulatorSetGoal)
+LuaC(0x6482E0, CSlideManipulatorBeenDestroyed)
+LuaC(0x649440, CreateStorageManip)
+LuaC(0x64AB40, CreateThrustController)
+LuaC(0x64AD10, CThrustManipulatorSetThrustingParam)
+LuaC(0x656450, CreateBeamEmitter)
+LuaC(0x656650, CreateBeamEmitterOnEntity)
+LuaC(0x6568E0, CreateBeamEntityToEntity)
+LuaC(0x656C30, AttachBeamEntityToEntity)
+LuaC(0x656F80, AttachBeamToEntity)
+LuaC(0x66D860, IEffectSetBeamParam)
+LuaC(0x66D940, IEffectSetEmitterParam)
+LuaC(0x66DA20, IEffectScaleEmitter)
+LuaC(0x66DBA0, IEffectResizeEmitterCurve)
+LuaC(0x66DE70, IEffectSetEmitterCurveParam)
+LuaC(0x66E150, IEffectOffsetEmitter)
+LuaC(0x66E3A0, IEffectDestroy)
+LuaC(0x66E4D0, CDecalHandleDestroy)
+LuaC(0x66E610, CreateDecal)
+LuaC(0x66EC10, CreateSplat)
+LuaC(0x66F1A0, CreateSplatOnBone)
+LuaC(0x66F6E0, CreateEmitterAtEntity)
+LuaC(0x66F8A0, CreateEmitterOnEntity)
+LuaC(0x66FA60, CreateLightParticle)
+LuaC(0x66FE20, CreateLightParticleIntel)
+LuaC(0x670240, CreateAttachedEmitter)
+LuaC(0x670420, CreateTrail)
+LuaC(0x670600, CreateAttachedBeam)
+LuaC(0x6708C0, CreateBeamToEntityBone)
+LuaC(0x670CF0, CreateEmitterAtBone)
+LuaC(0x673BC0, CollisionBeamEntity__init)
+LuaC(0x673F60, CollisionBeamEntitySetBeamFx)
+LuaC(0x674120, CollisionBeamEntityDisable)
+LuaC(0x674250, CollisionBeamEntityEnable)
+LuaC(0x674380, CollisionBeamEntityIsEnabled)
+LuaC(0x6744C0, CollisionBeamEntityGetLauncher)
+LuaC(0x68A090, EntityCreateProjectile)
+LuaC(0x68A670, EntityCreateProjectileAtBone)
+LuaC(0x68AAF0, EntityShakeCamera)
+LuaC(0x68AD60, EntityGetAIBrain)
+LuaC(0x68AF30, EntityGetBlueprint)
+LuaC(0x68B0B0, GetBlueprint)
+LuaC(0x68B210, EntityGetArmy)
+LuaC(0x68B370, EntityGetBoneDirection)
+LuaC(0x68B570, EntityIsValidBone)
+LuaC(0x68B7F0, EntityGetBoneCount)
+LuaC(0x68B940, EntityGetBoneName)
+LuaC(0x68BAE0, EntityRequestRefreshUI)
+LuaC(0x68BC10, EntityGetEntityId)
+LuaC(0x68BD90, EntityAttachTo)
+LuaC(0x68C010, EntityAttachBoneTo)
+LuaC(0x68C2B0, EntitySetParentOffset)
+LuaC(0x68C4A0, EntityDetachFrom)
+LuaC(0x68C650, EntityDetachAll)
+LuaC(0x68C8B0, EntityGetParent)
+LuaC(0x68CA00, EntityGetCollisionExtents)
+LuaC(0x68CC30, EntityPlaySound)
+LuaC(0x68CDC0, EntitySetAmbientSound)
+LuaC(0x68CF60, EntityGetFractionComplete)
+LuaC(0x68D0A0, EntityAdjustHealth)
+LuaC(0x68D2B0, EntityGetHealth)
+LuaC(0x68D3F0, EntityGetMaxHealth)
+LuaC(0x68D530, EntitySetHealth)
+LuaC(0x68D710, EntitySetMaxHealth)
+LuaC(0x68D880, EntitySetVizToFocusPlayer)
+LuaC(0x68DA00, EntitySetVizToEnemies)
+LuaC(0x68DB80, EntitySetVizToAllies)
+LuaC(0x68DD00, EntitySetVizToNeutrals)
+LuaC(0x68DE80, EntityIsIntelEnabled)
+LuaC(0x68E050, EntityEnableIntel)
+LuaC(0x68E2F0, EntityDisableIntel)
+LuaC(0x68E590, EntitySetIntelRadius)
+LuaC(0x68E7D0, EntityGetIntelRadius)
+LuaC(0x68E9A0, EntityInitIntel)
+LuaC(0x68ED50, EntityAddShooter)
+LuaC(0x68EEC0, EntityRemoveShooter)
+LuaC(0x68F030, EntityReachedMaxShooters)
+LuaC(0x68F180, EntitySetCollisionShape)
+LuaC(0x68F500, EntityGetOrientation)
+LuaC(0x68F660, EntitySetOrientation)
+LuaC(0x68F860, EntityGetHeading)
+LuaC(0x68FA10, EntitySetPosition)
+LuaC(0x68FC10, EntityGetPosition)
+LuaC(0x68FEE0, EntityGetPositionXYZ)
+LuaC(0x6900E0, EntityAddLocalImpulse)
+LuaC(0x6903E0, EntityAddWorldImpulse)
+LuaC(0x690790, EntitySetMesh)
+LuaC(0x6909F0, EntitySetDrawScale)
+LuaC(0x690B90, EntityGetScale)
+LuaC(0x690D20, EntitySetScale)
+LuaC(0x691030, EntityAddManualScroller)
+LuaC(0x691260, EntityAddThreadScroller)
+LuaC(0x691480, EntityAddPingPongScroller)
+LuaC(0x691840, EntityRemoveScroller)
+LuaC(0x691A00, EntityDestroy)
+LuaC(0x691B40, EntityBeenDestroyed)
+LuaC(0x691C90, EntityKill)
+LuaC(0x6922B0, _c_CreateEntity)
+LuaC(0x695580, EntityFallDown)
+LuaC(0x695720, MotorFallDownWhack)
+LuaC(0x6969E0, EntitySinkAway)
+LuaC(0x6A13B0, ProjectileGetLauncher)
+LuaC(0x6A1510, ProjectileGetTrackingTarget)
+LuaC(0x6A1680, ProjectileGetCurrentTargetPosition)
+LuaC(0x6A1820, ProjectileSetNewTarget)
+LuaC(0x6A19C0, ProjectileSetNewTargetGround)
+LuaC(0x6A1B90, ProjectileSetLifetime)
+LuaC(0x6A1D30, ProjectileSetDamage)
+LuaC(0x6A1F10, ProjectileSetMaxSpeed)
+LuaC(0x6A2090, ProjectileSetAcceleration)
+LuaC(0x6A2210, ProjectileSetBallisticAcceleration)
+LuaC(0x6A2470, ProjectileSetDestroyOnWater)
+LuaC(0x6A25B0, ProjectileSetTurnRate)
+LuaC(0x6A2730, ProjectileGetCurrentSpeed)
+LuaC(0x6A2890, ProjectileGetVelocity)
+LuaC(0x6A2A10, ProjectileSetVelocity)
+LuaC(0x6A2CF0, ProjectileSetScaleVelocity)
+LuaC(0x6A2F60, ProjectileSetLocalAngularVelocity)
+LuaC(0x6A3170, ProjectileSetCollision)
+LuaC(0x6A32E0, ProjectileSetCollideSurface)
+LuaC(0x6A3430, ProjectileSetCollideEntity)
+LuaC(0x6A3580, ProjectileStayUnderwater)
+LuaC(0x6A36D0, ProjectileTrackTarget)
+LuaC(0x6A3820, ProjectileSetStayUpright)
+LuaC(0x6A3960, ProjectileSetVelocityAlign)
+LuaC(0x6A3AA0, ProjectileCreateChildProjectile)
+LuaC(0x6A3CF0, ProjectileSetVelocityRandomUpVector)
+LuaC(0x6A3F00, ProjectileChangeMaxZigZag)
+LuaC(0x6A4070, ProjectileChangeZigZagFrequency)
+LuaC(0x6A41E0, ProjectileChangeDetonateAboveHeight)
+LuaC(0x6A4350, ProjectileChangeDetonateBelowHeight)
+LuaC(0x6C3AB0, UnitGetUnitId)
+LuaC(0x6C3C00, UnitSetCreator)
+LuaC(0x6C3D80, UnitGetCargo)
+LuaC(0x6C3F80, UnitAlterArmor)
+LuaC(0x6C4180, UnitGetArmorMult)
+LuaC(0x6C4370, UnitClearFocusEntity)
+LuaC(0x6C44E0, UnitSetFocusEntity)
+LuaC(0x6C4670, UnitGetFocusUnit)
+LuaC(0x6C4800, UnitGetWeapon)
+LuaC(0x6C49F0, UnitGetWeaponCount)
+LuaC(0x6C4B40, UnitGetTargetEntity)
+LuaC(0x6C4CA0, UnitGetHealth)
+LuaC(0x6C4E10, UnitGetAttacker)
+LuaC(0x6C4F60, UnitEnableManipulators)
+LuaC(0x6C5170, UnitKillManipulator)
+LuaC(0x6C52E0, UnitKillManipulators)
+LuaC(0x6C54B0, UnitScaleGetBuiltEmitter)
+LuaC(0x6C5780, UnitSetStrategicUnderlay)
+LuaC(0x6C5900, UnitIsUnitState)
+LuaC(0x6C5AA0, UnitIsIdleState)
+LuaC(0x6C5C10, UnitIsStunned)
+LuaC(0x6C5D60, UnitIsBeingBuilt)
+LuaC(0x6C5EB0, UnitIsPaused)
+LuaC(0x6C6000, UnitSetPaused)
+LuaC(0x6C6140, UnitSetConsumptionActive)
+LuaC(0x6C6290, UnitSetProductionActive)
+LuaC(0x6C63F0, UnitSetBusy)
+LuaC(0x6C6570, UnitSetBlockCommandQueue)
+LuaC(0x6C66F0, UnitSetImmobile)
+LuaC(0x6C6860, UnitSetStunned)
+LuaC(0x6C69D0, UnitSetUnSelectable)
+LuaC(0x6C6B40, UnitSetDoNotTarget)
+LuaC(0x6C6CB0, UnitSetUnitState)
+LuaC(0x6C6E80, UnitStopSiloBuild)
+LuaC(0x6C6FB0, UnitSetIsValidTarget)
+LuaC(0x6C70F0, UnitIsValidTarget)
+LuaC(0x6C7230, UnitGetNumBuildOrders)
+LuaC(0x6C7470, UnitCalculateWorldPositionFromRelative)
+LuaC(0x6C7690, UnitGetScriptBit)
+LuaC(0x6C7870, UnitSetScriptBit)
+LuaC(0x6C7A30, UnitToggleScriptBit)
+LuaC(0x6C7BD0, UnitToggleFireState)
+LuaC(0x6C7D10, UnitSetFireState)
+LuaC(0x6C7EA0, UnitGetFireState)
+LuaC(0x6C7FE0, UnitSetAutoMode)
+LuaC(0x6C8120, UnitAddBuildRestriction)
+LuaC(0x6C8290, UnitRemoveBuildRestriction)
+LuaC(0x6C8400, UnitRestoreBuildRestrictions)
+LuaC(0x6C8560, UnitAddCommandCap)
+LuaC(0x6C8700, UnitRemoveCommandCap)
+LuaC(0x6C88A0, UnitRestoreCommandCaps)
+LuaC(0x6C89E0, UnitTestCommandCaps)
+LuaC(0x6C8B80, UnitAddToggleCap)
+LuaC(0x6C8D20, UnitRemoveToggleCap)
+LuaC(0x6C8EC0, UnitRestoreToggleCaps)
+LuaC(0x6C9000, UnitTestToggleCaps)
+LuaC(0x6C91A0, UnitSetRegenRate)
+LuaC(0x6C9330, UnitRevertRegenRate)
+LuaC(0x6C9460, UnitSetReclaimable)
+LuaC(0x6C95B0, UnitSetCapturable)
+LuaC(0x6C9700, UnitIsCapturable)
+LuaC(0x6C9840, UnitSetOverchargePaused)
+LuaC(0x6C9980, UnitIsOverchargePaused)
+LuaC(0x6C9AC0, UnitSetBuildRate)
+LuaC(0x6C9C50, UnitGetBuildRate)
+LuaC(0x6C9DA0, UnitSetConsumptionPerSecondEnergy)
+LuaC(0x6C9F40, UnitSetConsumptionPerSecondMass)
+LuaC(0x6CA0E0, UnitSetProductionPerSecondEnergy)
+LuaC(0x6CA270, UnitSetProductionPerSecondMass)
+LuaC(0x6CA400, UnitGetConsumptionPerSecondEnergy)
+LuaC(0x6CA550, UnitGetConsumptionPerSecondMass)
+LuaC(0x6CA6A0, UnitGetProductionPerSecondEnergy)
+LuaC(0x6CA7F0, UnitGetProductionPerSecondMass)
+LuaC(0x6CA940, UnitGetResourceConsumed)
+LuaC(0x6CAA80, UnitSetElevation)
+LuaC(0x6CABF0, UnitRevertElevation)
+LuaC(0x6CAD20, UnitSetSpeedMult)
+LuaC(0x6CAE90, UnitSetAccMult)
+LuaC(0x6CB000, UnitSetTurnMult)
+LuaC(0x6CB170, UnitSetBreakOffTriggerMult)
+LuaC(0x6CB2E0, UnitSetBreakOffDistanceMult)
+LuaC(0x6CB450, UnitRevertCollisionShape)
+LuaC(0x6CB580, UnitRecoilImpulse)
+LuaC(0x6CB790, UnitGetCurrentLayer)
+LuaC(0x6CB8E0, UnitCanPathTo)
+LuaC(0x6CBCF0, UnitCanPathToRect)
+LuaC(0x6CC040, UnitIsMobile)
+LuaC(0x6CC180, UnitIsMoving)
+LuaC(0x6CC2D0, UnitGetNavigator)
+LuaC(0x6CC420, UnitGetVelocity)
+LuaC(0x6CC5A0, UnitGetStat)
+LuaC(0x6CC8E0, UnitSetStat)
+LuaC(0x6CCB40, UnitSetWorkProgress)
+LuaC(0x6CCCB0, UnitGetWorkProgress)
+LuaC(0x6CCDF0, NotifyUpgrade)
+LuaC(0x6CD300, UnitGetGuardedUnit)
+LuaC(0x6CD460, UnitGetGuards)
+LuaC(0x6CD660, UnitGetTransportFerryBeacon)
+LuaC(0x6CD7E0, UnitHasValidTeleportDest)
+LuaC(0x6CD950, UnitAddUnitToStorage)
+LuaC(0x6CDB00, UnitSetCustomName)
+LuaC(0x6CDC90, UnitHasMeleeSpaceAroundTarget)
+LuaC(0x6CDED0, UnitMeleeWarpAdjacentToTarget)
+LuaC(0x6CE200, UnitGetCommandQueue)
+LuaC(0x6CE420, UnitPrintCommandQueue)
+LuaC(0x6CEA60, UnitGetCurrentMoveLocation)
+LuaC(0x6CECC0, UnitGiveNukeSiloAmmo)
+LuaC(0x6CEE10, UnitRemoveNukeSiloAmmo)
+LuaC(0x6CEF60, UnitGetNukeSiloAmmoCount)
+LuaC(0x6CF0D0, UnitGiveTacticalSiloAmmo)
+LuaC(0x6CF220, UnitRemoveTacticalSiloAmmo)
+LuaC(0x6CF370, UnitGetTacticalSiloAmmoCount)
+LuaC(0x6CF4E0, CreateUnit)
+LuaC(0x6CFAF0, CreateUnitHPR)
+LuaC(0x6D0000, CreateUnit2)
+LuaC(0x6D03E0, UnitCanBuild)
+LuaC(0x6D05D0, UnitGetRallyPoint)
+LuaC(0x6D0770, UnitGetFuelUseTime)
+LuaC(0x6D08D0, UnitSetFuelUseTime)
+LuaC(0x6D0A50, UnitGetFuelRatio)
+LuaC(0x6D0BB0, UnitSetFuelRatio)
+LuaC(0x6D0D30, UnitSetShieldRatio)
+LuaC(0x6D0EA0, UnitGetShieldRatio)
+LuaC(0x6D0FE0, UnitGetBlip)
+LuaC(0x6D1170, UnitTransportHasSpaceFor)
+LuaC(0x6D1320, UnitTransportHasAvailableStorage)
+LuaC(0x6D1470, UnitShowBone)
+LuaC(0x6D1610, UnitHideBone)
+LuaC(0x6D7BD0, UnitWeaponPlaySound)
+LuaC(0x6D7D70, UnitWeaponSetEnabled)
+LuaC(0x6D7F00, UnitWeaponSetTargetEntity)
+LuaC(0x6D80C0, UnitWeaponSetTargetGround)
+LuaC(0x6D8320, UnitWeaponResetTarget)
+LuaC(0x6D8490, UnitWeaponCreateProjectile)
+LuaC(0x6D8610, UnitWeaponDoInstaHit)
+LuaC(0x6D8A60, UnitWeaponGetBlueprint)
+LuaC(0x6D8C10, UnitWeaponGetProjectileBlueprint)
+LuaC(0x6D8D70, UnitWeaponWeaponHasTarget)
+LuaC(0x6D8EB0, UnitWeaponFireWeapon)
+LuaC(0x6D8FE0, UnitWeaponSetFireControl)
+LuaC(0x6D9150, UnitWeaponIsFireControl)
+LuaC(0x6D9310, UnitWeaponGetCurrentTarget)
+LuaC(0x6D9490, UnitWeaponGetCurrentTargetPos)
+LuaC(0x6D9630, UnitWeaponCanFire)
+LuaC(0x6D97C0, UnitWeaponChangeFiringTolerance)
+LuaC(0x6D9920, UnitWeaponChangeRateOfFire)
+LuaC(0x6D9A80, UnitWeaponChangeMinRadius)
+LuaC(0x6D9C00, UnitWeaponChangeMaxRadius)
+LuaC(0x6D9D80, UnitWeaponChangeMaxHeightDiff)
+LuaC(0x6D9EE0, UnitWeaponChangeDamageType)
+LuaC(0x6DA070, UnitWeaponChangeDamageRadius)
+LuaC(0x6DA1E0, UnitWeaponChangeDamage)
+LuaC(0x6DA350, UnitWeaponSetFireTargetLayerCaps)
+LuaC(0x6DA4E0, UnitWeaponSetTargetingPriorities)
+LuaC(0x6DA6B0, UnitWeaponGetFiringRandomness)
+LuaC(0x6DA7F0, UnitWeaponSetFiringRandomness)
+LuaC(0x6DA960, UnitWeaponGetFireClockPct)
+LuaC(0x6DAAE0, UnitWeaponChangeProjectileBlueprint)
+LuaC(0x6DAC80, UnitWeaponTransferTarget)
+LuaC(0x6DADF0, UnitWeaponBeenDestroyed)
+LuaC(0x6F17B0, IsCommandDone)
+LuaC(0x6F18E0, IssueClearCommands)
+LuaC(0x6F1A40, IssueStop)
+LuaC(0x6F1BE0, IssuePause)
+LuaC(0x6F1D80, IssueOverCharge)
+LuaC(0x6F2030, IssueDive)
+LuaC(0x6F2250, IssueFactoryRallyPoint)
+LuaC(0x6F2510, IssueClearFactoryCommands)
+LuaC(0x6F2660, IssueMove)
+LuaC(0x6F2960, IssueMoveOffFactory)
+LuaC(0x6F2C70, IssueFormMove)
+LuaC(0x6F30D0, IssueGuard)
+LuaC(0x6F33A0, IssueFactoryAssist)
+LuaC(0x6F3650, IssueAttack)
+LuaC(0x6F3910, CoordinateAttacks)
+LuaC(0x6F3B60, IssueFormAttack)
+LuaC(0x6F3F40, IssueSiloBuildTactical)
+LuaC(0x6F40A0, IssueSiloBuildNuke)
+LuaC(0x6F4200, IssueNuke)
+LuaC(0x6F4490, IssueTactical)
+LuaC(0x6F4720, IssueTeleport)
+LuaC(0x6F49B0, IssuePatrol)
+LuaC(0x6F4C80, IssueFormPatrol)
+LuaC(0x6F50B0, IssueAggressiveMove)
+LuaC(0x6F53C0, IssueFormAggressiveMove)
+LuaC(0x6F5820, IssueFerry)
+LuaC(0x6F5AF0, IssueBuildMobile)
+LuaC(0x6F6060, IssueRepair)
+LuaC(0x6F6330, IssueSacrifice)
+LuaC(0x6F6600, IssueUpgrade)
+LuaC(0x6F67A0, IssueScript)
+LuaC(0x6F6960, IssueReclaim)
+LuaC(0x6F6C20, IssueCapture)
+LuaC(0x6F6EE0, IssueKillSelf)
+LuaC(0x6F7060, IssueDestroySelf)
+LuaC(0x6F71E0, IssueTransportLoad)
+LuaC(0x6F7530, IssueTransportUnload)
+LuaC(0x6F77E0, IssueTeleportToBeacon)
+LuaC(0x6F7A70, IssueTransportUnloadSpecific)
+LuaC(0x6F7EA0, IssueBuildFactory)
+LuaC(0x6FBF90, CreatePropHPR)
+LuaC(0x6FC3B0, CreateProp)
+LuaC(0x6FC590, EntityCreatePropAtBone)
+LuaC(0x6FC830, SplitProp)
+LuaC(0x6FCA80, EntityPushOver)
+LuaC(0x6FCF40, PropAddBoundedProp)
+LuaC(0x7080B0, ShouldCreateInitialArmyUnits)
+LuaC(0x7081A0, ListArmies)
+LuaC(0x708310, GetArmyBrain)
+LuaC(0x708460, SetArmyStart)
+LuaC(0x7085E0, GenerateArmyStart)
+LuaC(0x7086D0, SetArmyPlans)
+LuaC(0x708870, InitializeArmyAI)
+LuaC(0x708970, ArmyInitializePrebuiltUnits)
+LuaC(0x708A70, ArmyGetHandicap)
+LuaC(0x708B90, SetArmyEconomy)
+LuaC(0x708D60, GetArmyUnitCostTotal)
+LuaC(0x708E60, GetArmyUnitCap)
+LuaC(0x708F70, SetArmyUnitCap)
+LuaC(0x7090A0, SetIgnoreArmyUnitCap)
+LuaC(0x7091B0, SetIgnorePlayableRect)
+LuaC(0x7092C0, CreateInitialArmyUnit)
+LuaC(0x709590, SetAlliance)
+LuaC(0x709720, SetAllianceOneWay)
+LuaC(0x7098A0, SetAlliedVictory)
+LuaC(0x7099C0, IsAlly)
+LuaC(0x709AF0, IsEnemy)
+LuaC(0x709C20, IsNeutral)
+LuaC(0x709D50, ArmyIsCivilian)
+LuaC(0x709E50, SetArmyColorIndex)
+LuaC(0x709FB0, SetArmyFactionIndex)
+LuaC(0x70A100, SetArmyAIPersonality)
+LuaC(0x70A2A0, SetArmyColor)
+LuaC(0x70A5E0, SetArmyShowScore)
+LuaC(0x70A6E0, AddBuildRestriction)
+LuaC(0x70A800, RemoveBuildRestriction)
+LuaC(0x70A920, OkayToMessWithArmy)
+LuaC(0x70AA60, ArmyIsOutOfGame)
+LuaC(0x70AB60, SetArmyOutOfGame)
+LuaC(0x72B810, CPlatoonIsOpponentAIRunning)
+LuaC(0x72B970, CPlatoonGetPersonality)
+LuaC(0x72BAD0, CPlatoonGetBrain)
+LuaC(0x72BC10, CPlatoonGetFactionIndex)
+LuaC(0x72BD60, CPlatoonUniquelyNamePlatoon)
+LuaC(0x72BEF0, CPlatoonGetPlatoonUniqueName)
+LuaC(0x72C040, CPlatoonGetAIPlan)
+LuaC(0x72C190, CPlatoonSwitchAIPlan)
+LuaC(0x72C300, CPlatoonGetPlatoonPosition)
+LuaC(0x72C5C0, CPlatoonGetSquadPosition)
+LuaC(0x72C7D0, CPlatoonGetSquadUnits)
+LuaC(0x72C9B0, CPlatoonGetPlatoonUnits)
+LuaC(0x72CC00, CPlatoonCanConsiderFormingPlatoon)
+LuaC(0x72CE70, CPlatoonCanFormPlatoon)
+LuaC(0x72D870, CPlatoonFormPlatoon)
+LuaC(0x72E240, CPlatoonIsAttacking)
+LuaC(0x72E3E0, CPlatoonIsMoving)
+LuaC(0x72E580, CPlatoonIsPatrolling)
+LuaC(0x72E720, CPlatoonIsFerrying)
+LuaC(0x72E8C0, CPlatoonSetPrioritizedTargetList)
+LuaC(0x72EB30, CPlatoonFindPrioritizedUnit)
+LuaC(0x72EDE0, CPlatoonFindClosestUnit)
+LuaC(0x72F020, CPlatoonFindClosestUnitToBase)
+LuaC(0x72F290, CPlatoonFindFurthestUnit)
+LuaC(0x72F4D0, CPlatoonFindHighestValueUnit)
+LuaC(0x72F710, CPlatoonCanAttackTarget)
+LuaC(0x72F940, CPlatoonStop)
+LuaC(0x72FAE0, CPlatoonAttackTarget)
+LuaC(0x72FDB0, CPlatoonMoveToTarget)
+LuaC(0x7300A0, CPlatoonMoveToLocation)
+LuaC(0x7303A0, CPlatoonAggressiveMoveToLocation)
+LuaC(0x730680, CPlatoonFerryToLocation)
+LuaC(0x7308F0, CPlatoonLoadUnits)
+LuaC(0x730B10, CPlatoonUnloadUnitsAtLocation)
+LuaC(0x730DB0, CPlatoonUnloadAllAtLocation)
+LuaC(0x731020, CPlatoonGuardTarget)
+LuaC(0x7312F0, CPlatoonPatrol)
+LuaC(0x7314F0, CPlatoonDestroy)
+LuaC(0x731690, CPlatoonDisbandOnIdle)
+LuaC(0x7317C0, CPlatoonGetFerryBeacons)
+LuaC(0x7319B0, CPlatoonUseFerryBeacon)
+LuaC(0x731C30, CPlatoonUseTeleporter)
+LuaC(0x731F00, CPlatoonIsCommandsActive)
+LuaC(0x7320C0, CPlatoonSetPlatoonFormationOverride)
+LuaC(0x7322D0, CPlatoonGetPlatoonLifetimeStats)
+LuaC(0x7324A0, CPlatoonCalculatePlatoonThreat)
+LuaC(0x7328E0, CPlatoonCalculatePlatoonThreatAroundPosition)
+LuaC(0x732E40, CPlatoonPlatoonCategoryCountAroundPosition)
+LuaC(0x7331C0, CPlatoonPlatoonCategoryCount)
+LuaC(0x738630, CDamageGetInstigator)
+LuaC(0x738790, CDamageSetInstigator)
+LuaC(0x738900, CDamageGetTarget)
+LuaC(0x738A60, CDamageSetTarget)
+LuaC(0x738BD0, Damage)
+LuaC(0x738F40, DamageArea)
+LuaC(0x7392C0, DamageRing)
+LuaC(0x7396E0, MetaImpact)
+LuaC(0x74B570, Sim_print)
+LuaC(0x74B620, CheatsEnabled)
+LuaC(0x74B710, GetCurrentCommandSource)
+LuaC(0x74B830, EndGame)
+LuaC(0x74B900, IsGameOver)
+LuaC(0x74B9F0, GetEntityById)
+LuaC(0x74BB40, GetUnitById)
+LuaC(0x758F90, Random)
+LuaC(0x759190, FlattenMapRect)
+LuaC(0x7593D0, SelectedUnit)
+LuaC(0x7594C0, SimConExecute)
+LuaC(0x759810, ParseEntityCategory)
+LuaC(0x759990, EntityCategoryContains)
+LuaC(0x759BD0, EntityCategoryFilterDown)
+LuaC(0x759EF0, EntityCategoryCount)
+LuaC(0x75A1D0, EntityCategoryCountAroundPosition)
+LuaC(0x75A580, GenerateRandomOrientation)
+LuaC(0x75A770, GetGameTimeSeconds)
+LuaC(0x75A860, GetGameTick)
+LuaC(0x75A950, GetSystemTimeSecondsOnlyForProfileUse)
+LuaC(0x75AA30, GetEntitiesInRect)
+LuaC(0x75AE00, GetUnitsInRect)
+LuaC(0x75B200, GetReclaimablesInRect)
+LuaC(0x75B5D0, ChangeUnitArmy)
+LuaC(0x75B940, Warp)
+LuaC(0x75BBE0, GetMapSize)
+LuaC(0x75BCF0, GetTerrainHeight)
+LuaC(0x75BE70, GetSurfaceHeight)
+LuaC(0x75C030, GetTerrainTypeOffset)
+LuaC(0x75C1B0, GetTerrainType)
+LuaC(0x75C3B0, SetTerrainType)
+LuaC(0x75C5D0, SetTerrainTypeRect)
+LuaC(0x75C7B0, SetPlayableRect)
+LuaC(0x75C9D0, FlushIntelInRect)
+LuaC(0x75CC20, GetUnitBlueprintByName)
+LuaC(0x75CDD0, GetFocusArmy)
+LuaC(0x75CEC0, AudioSetLanguage)
+LuaC(0x75CF90, HasLocalizedVO)
+LuaC(0x75D060, TryCopyPose)
+LuaC(0x75D970, SetArmyStatsSyncArmy)
+LuaC(0x75DA80, SubmitXMLArmyStats)
+LuaC(0x75DB80, DrawLine)
+LuaC(0x75DDA0, DrawLinePop)
+LuaC(0x75E0B0, DrawCircle)
+LuaC(0x75E2A0, DebugGetSelection)
+LuaC(0x75E4E0, IsEntity)
+LuaC(0x75E620, IsUnit)
+LuaC(0x75E780, IsProp)
+LuaC(0x75E8E0, IsBlip)
+LuaC(0x75EA40, IsProjectile)
+LuaC(0x75EBA0, IsCollisionBeam)
+LuaC(0x761570, PlayLoop)
+LuaC(0x761700, StopLoop)
+LuaC(0x775630, CreateEconomyEvent)
+LuaC(0x775910, RemoveEconomyEvent)
+LuaC(0x775A40, EconomyEventIsDone)
+LuaC(0x776A20, _c_CreateShield)
+LuaC(0x780A70, GetTextureDimensions)
+LuaC(0x780D20, InternalCreateBitmap)
+LuaC(0x780ED0, CMauiBitmapSetNewTexture)
+LuaC(0x781440, CMauiBitmapInternalSetSolidColor)
+LuaC(0x781690, CMauiBitmapSetUV)
+LuaC(0x781950, CMauiBitmapUseAlphaHitTest)
+LuaC(0x781AA0, CMauiBitmapSetTiled)
+LuaC(0x781BF0, CMauiBitmapLoop)
+LuaC(0x781D40, CMauiBitmapPlay)
+LuaC(0x781EA0, CMauiBitmapStop)
+LuaC(0x782000, CMauiBitmapSetFrame)
+LuaC(0x782180, CMauiBitmapGetFrame)
+LuaC(0x7822C0, CMauiBitmapGetNumFrames)
+LuaC(0x782420, CMauiBitmapSetFrameRate)
+LuaC(0x7825A0, CMauiBitmapSetForwardPattern)
+LuaC(0x7826E0, CMauiBitmapSetBackwardPattern)
+LuaC(0x782820, CMauiBitmapSetPingPongPattern)
+LuaC(0x782960, CMauiBitmapSetLoopPingPongPattern)
+LuaC(0x782AA0, CMauiBitmapSetFramePattern)
+LuaC(0x782CE0, CMauiBitmapShareTextures)
+LuaC(0x7857B0, InternalCreateBorder)
+LuaC(0x785960, CMauiBorderSetNewTextures)
+LuaC(0x785FA0, CMauiBorderSetSolidColor)
+LuaC(0x787A50, CMauiControlDestroy)
+LuaC(0x787BA0, CMauiControlGetParent)
+LuaC(0x787CF0, CMauiControlClearChildren)
+LuaC(0x787E30, CMauiControlSetParent)
+LuaC(0x787FB0, CMauiControlDisableHitTest)
+LuaC(0x788130, CMauiControlEnableHitTest)
+LuaC(0x7882B0, CMauiControlIsHitTestDisabled)
+LuaC(0x7883F0, CMauiControlHide)
+LuaC(0x788520, CMauiControlShow)
+LuaC(0x788650, CMauiControlSetHidden)
+LuaC(0x788790, CMauiControlIsHidden)
+LuaC(0x7888D0, CMauiControlGetRenderPass)
+LuaC(0x788A10, CMauiControlSetRenderPass)
+LuaC(0x788B90, CMauiControlGetName)
+LuaC(0x788D00, CMauiControlSetName)
+LuaC(0x788E90, CMauiControlDump)
+LuaC(0x788FC0, CMauiControlGetCurrentFocusControl)
+LuaC(0x7890B0, CMauiControlAcquireKeyboardFocus)
+LuaC(0x789210, CMauiControlAbandonKeyboardFocus)
+LuaC(0x789350, CMauiControlNeedsFrameUpdate)
+LuaC(0x789490, CMauiControlSetNeedsFrameUpdate)
+LuaC(0x7895D0, CMauiControlGetRootFrame)
+LuaC(0x789710, CMauiControlSetAlpha)
+LuaC(0x789A30, CMauiControlGetAlpha)
+LuaC(0x789B70, CMauiControlApplyFunction)
+LuaC(0x789CD0, CMauiControlHitTest)
+LuaC(0x78CFD0, _c_CreateCursor)
+LuaC(0x78D130, CMauiCursorSetNewTexture)
+LuaC(0x78D390, CMauiCursorSetDefaultTexture)
+LuaC(0x78D5A0, CMauiCursorResetToDefault)
+LuaC(0x78D6C0, CMauiCursorHide)
+LuaC(0x78D7F0, CMauiCursorShow)
+LuaC(0x78DF80, CMauiLuaDraggerDestroy)
+LuaC(0x78E0B0, InternalCreateDragger)
+LuaC(0x78E210, PostDragger)
+LuaC(0x791FF0, InternalCreateEdit)
+LuaC(0x7921A0, CMauiEditSetNewFont)
+LuaC(0x7923C0, CMauiEditSetNewForegroundColor)
+LuaC(0x792530, CMauiEditGetForegroundColor)
+LuaC(0x792690, CMauiEditSetNewBackgroundColor)
+LuaC(0x792810, CMauiEditGetBackgroundColor)
+LuaC(0x792970, CMauiEditShowBackground)
+LuaC(0x792AC0, CMauiEditIsBackgroundVisible)
+LuaC(0x792C00, CMauiEditClearText)
+LuaC(0x792D30, CMauiEditSetText)
+LuaC(0x792F00, CMauiEditGetText)
+LuaC(0x793070, CMauiEditSetMaxChars)
+LuaC(0x793200, CMauiEditGetMaxChars)
+LuaC(0x793340, CMauiEditSetCaretPosition)
+LuaC(0x7934C0, CMauiEditGetCaretPosition)
+LuaC(0x793600, CMauiEditShowCaret)
+LuaC(0x793750, CMauiEditIsCaretVisible)
+LuaC(0x793890, CMauiEditSetNewCaretColor)
+LuaC(0x793A00, CMauiEditGetCaretColor)
+LuaC(0x793B60, CMauiEditSetCaretCycle)
+LuaC(0x793D70, CMauiEditIsEnabled)
+LuaC(0x793EB0, CMauiEditEnableInput)
+LuaC(0x793FF0, CMauiEditDisableInput)
+LuaC(0x794140, CMauiEditSetNewHighlightForegroundColor)
+LuaC(0x7942B0, CMauiEditGetHighlightForegroundColor)
+LuaC(0x794410, CMauiEditSetNewHighlightBackgroundColor)
+LuaC(0x794580, CMauiEditGetHighlightBackgroundColor)
+LuaC(0x7946E0, CMauiEditGetFontHeight)
+LuaC(0x794840, CMauiEditAcquireFocus)
+LuaC(0x794990, CMauiEditAbandonFocus)
+LuaC(0x794AD0, CMauiEditSetDropShadow)
+LuaC(0x794C20, CMauiEditGetStringAdvance)
+LuaC(0x796790, InternalCreateFrame)
+LuaC(0x796900, CMauiFrameGetTopmostDepth)
+LuaC(0x796A50, CMauiFrameGetTargetHead)
+LuaC(0x796B90, CMauiFrameSetTargetHead)
+LuaC(0x797310, InternalCreateGroup)
+LuaC(0x797920, InternalCreateHistogram)
+LuaC(0x797AD0, CMauiHistogramSetXIncrement)
+LuaC(0x797C50, CMauiHistogramSetYIncrement)
+LuaC(0x797DD0, CMauiHistogramSetData)
+LuaC(0x79A960, InternalCreateItemList)
+LuaC(0x79AB10, CMauiItemListSetNewFont)
+LuaC(0x79AD30, CMauiItemListSetNewColors)
+LuaC(0x79B040, CMauiItemListGetItem)
+LuaC(0x79B1E0, CMauiItemListAddItem)
+LuaC(0x79B370, CMauiItemListModifyItem)
+LuaC(0x79B560, CMauiItemListDeleteItem)
+LuaC(0x79B6E0, CMauiItemListDeleteAllItems)
+LuaC(0x79B840, CMauiItemListGetSelection)
+LuaC(0x79B980, CMauiItemListSetSelection)
+LuaC(0x79BB40, CMauiItemListGetItemCount)
+LuaC(0x79BCB0, CMauiItemListEmpty)
+LuaC(0x79BE10, CMauiItemListScrollToTop)
+LuaC(0x79BF40, CMauiItemListScrollToBottom)
+LuaC(0x79C070, CMauiItemListShowItem)
+LuaC(0x79C200, CMauiItemListGetRowHeight)
+LuaC(0x79C350, CMauiItemListGetStringAdvance)
+LuaC(0x79C4E0, CMauiItemListShowMouseoverItem)
+LuaC(0x79C620, CMauiItemListShowSelection)
+LuaC(0x79C760, CMauiItemListNeedsScrollBar)
+LuaC(0x79D7A0, IsKeyDown)
+LuaC(0x79D8D0, KeycodeMauiToMSW)
+LuaC(0x79D9F0, KeycodeMSWToMaui)
+LuaC(0x79E590, InternalCreateMesh)
+LuaC(0x79E740, CMauiMeshSetMesh)
+LuaC(0x79E8B0, CMauiMeshSetOrientation)
+LuaC(0x79F540, InternalCreateMovie)
+LuaC(0x79F6F0, CMauiMovieInternalSet)
+LuaC(0x79F870, CMauiMovieLoop)
+LuaC(0x79F9C0, CMauiMoviePlay)
+LuaC(0x79FB20, CMauiMovieStop)
+LuaC(0x79FC70, CMauiMovieIsLoaded)
+LuaC(0x79FDD0, CMauiMovieGetNumFrames)
+LuaC(0x79FF20, CMauiMovieGetFrameRate)
+LuaC(0x7A1590, InternalCreateScrollbar)
+LuaC(0x7A17A0, CMauiScrollbarSetScrollable)
+LuaC(0x7A1920, CMauiScrollbarSetNewTextures)
+LuaC(0x7A2080, CMauiScrollbarDoScrollLines)
+LuaC(0x7A2220, CMauiScrollbarDoScrollPages)
+LuaC(0x7A3340, InternalCreateText)
+LuaC(0x7A34F0, CMauiTextSetNewFont)
+LuaC(0x7A3710, CMauiTextSetText)
+LuaC(0x7A3880, CMauiTextGetText)
+LuaC(0x7A39F0, CMauiTextSetNewColor)
+LuaC(0x7A3B60, CMauiTextSetDropShadow)
+LuaC(0x7A3CB0, CMauiTextSetCenteredHorizontally)
+LuaC(0x7A3E00, CMauiTextSetCenteredVertically)
+LuaC(0x7A3F50, CMauiTextGetStringAdvance)
+LuaC(0x7A40E0, CMauiTextSetNewClipToWidth)
+LuaC(0x7A5190, AnyInputCapture)
+LuaC(0x7A5280, GetInputCapture)
+LuaC(0x7A53B0, AddInputCapture)
+LuaC(0x7A54E0, RemoveInputCapture)
+LuaC(0x7AB080, GetCamera)
+LuaC(0x7AB1B0, CameraImplMoveToRegion)
+LuaC(0x7AB4E0, CameraImplSnapTo)
+LuaC(0x7AB6E0, CameraImplMoveTo)
+LuaC(0x7AB930, CameraImplReset)
+LuaC(0x7ABA60, CameraImplTrackEntities)
+LuaC(0x7ABE40, CameraImplTargetEntities)
+LuaC(0x7AC1C0, CameraImplNoseCam)
+LuaC(0x7AC500, CameraImplHoldRotation)
+LuaC(0x7AC630, CameraImplRevertRotation)
+LuaC(0x7AC760, CameraImplSetZoom)
+LuaC(0x7AC930, CameraImplSetTargetZoom)
+LuaC(0x7ACAA0, CameraImplSetAccMode)
+LuaC(0x7ACC40, CameraImplGetZoom)
+LuaC(0x7ACD80, CameraImplGetFocusPosition)
+LuaC(0x7ACEE0, CameraImplSaveSettings)
+LuaC(0x7AD0D0, CameraImplRestoreSettings)
+LuaC(0x7AD3D0, CameraImplGetTargetZoom)
+LuaC(0x7AD510, CameraImplGetMaxZoom)
+LuaC(0x7AD650, CameraImplGetMinZoom)
+LuaC(0x7AD720, CameraImplSetMaxZoomMult)
+LuaC(0x7AD890, CameraImplSpin)
+LuaC(0x7ADA90, CameraImplUseGameClock)
+LuaC(0x7ADBC0, CameraImplUseSystemClock)
+LuaC(0x7ADCF0, CameraImplEnableEaseInOut)
+LuaC(0x7ADE20, CameraImplDisableEaseInOut)
+LuaC(0x7B5170, CreateUnitAtMouse)
+LuaC(0x7B5E60, CPathDebuggerDestroy)
+LuaC(0x7B5F90, _c_CreatePathDebugger)
+LuaC(0x7B9DE0, GpgNetActive)
+LuaC(0x7B9EB0, GpgNetSend)
+LuaC(0x7BA2C0, LaunchGPGNet)
+LuaC(0x7C0060, InternalCreateDiscoveryService)
+LuaC(0x7C01C0, CDiscoveryServiceGetGameCount)
+LuaC(0x7C0310, CDiscoveryServiceReset)
+LuaC(0x7C0430, CDiscoveryServiceDestroy)
+LuaC(0x7C0CC0, InternalCreateLobby)
+LuaC(0x7C13E0, CLobbyDestroy)
+LuaC(0x7C1530, CLobbyMakeValidGameName)
+LuaC(0x7C18D0, CLobbyMakeValidPlayerName)
+LuaC(0x7C1C80, CLobbyHostGame)
+LuaC(0x7C1FA0, CLobbyJoinGame)
+LuaC(0x7C2350, CLobbyBroadcastData)
+LuaC(0x7C2650, CLobbySendData)
+LuaC(0x7C2B00, CLobbyGetPeers)
+LuaC(0x7C2C60, CLobbyGetPeer)
+LuaC(0x7C2DF0, CLobbyGetLocalPlayerName)
+LuaC(0x7C2F40, CLobbyGetLocalPlayerID)
+LuaC(0x7C3090, CLobbyIsHost)
+LuaC(0x7C31D0, CLobbyGetLocalPort)
+LuaC(0x7C3340, CLobbyEjectPeer)
+LuaC(0x7C34D0, CLobbyConnectToPeer)
+LuaC(0x7C3760, CLobbyDisconnectFromPeer)
+LuaC(0x7C50E0, CLobbyLaunchGame)
+LuaC(0x7C5360, CLobbyDebugDump)
+LuaC(0x7C82E0, ValidateIPAddress)
+LuaC(0x7FDAF0, AddBlinkyBox)
+LuaC(0x836360, SetCurrentFactoryForQueueDisplay)
+LuaC(0x8365B0, ClearCurrentFactoryForQueueDisplay)
+LuaC(0x8366C0, IncreaseBuildCountInQueue)
+LuaC(0x836900, DecreaseBuildCountInQueue)
+LuaC(0x83A190, IN_AddKeyMapTable)
+LuaC(0x83A2B0, IN_RemoveKeyMapTable)
+LuaC(0x83A3D0, IN_ClearKeyMap)
+LuaC(0x83DA90, ConTextMatches)
+LuaC(0x83F000, EngineStartSplashScreens)
+LuaC(0x83F0E0, EngineStartFrontEndUI)
+LuaC(0x83F1C0, ExitApplication)
+LuaC(0x83F2B0, ExitGame)
+LuaC(0x83F3A0, RestartSession)
+LuaC(0x83F470, GetFrame)
+LuaC(0x83F5D0, ClearFrame)
+LuaC(0x83F700, GetNumRootFrames)
+LuaC(0x83F800, GetEconomyTotals)
+LuaC(0x83FED0, GetResourceSharing)
+LuaC(0x83FFE0, GetSimTicksPerSecond)
+LuaC(0x8400B0, GetUnitCommandData)
+LuaC(0x840840, GetUnitCommandFromCommandCap)
+LuaC(0x8409F0, IssueDockCommand)
+LuaC(0x841530, IssueCommand)
+LuaC(0x841840, IssueUnitCommand)
+LuaC(0x841B90, IssueBlueprintCommand)
+LuaC(0x8420A0, GetCurrentUIState)
+LuaC(0x8428A0, GetRolloverInfo)
+LuaC(0x842BB0, GetMouseWorldPos)
+LuaC(0x842D10, GetMouseScreenPos)
+LuaC(0x842E60, UI_SetFocusArmy)
+LuaC(0x842FD0, UI_GetFocusArmy)
+LuaC(0x8430D0, IsObserver)
+LuaC(0x8431D0, GetGameTime)
+LuaC(0x843380, UI_GetGameTimeSeconds)
+LuaC(0x843480, GetSystemTime)
+LuaC(0x8435F0, GetSystemTimeSeconds)
+LuaC(0x8436D0, FormatTime)
+LuaC(0x8438A0, GetSimRate)
+LuaC(0x8439A0, GetArmiesTable)
+LuaC(0x843E50, GetArmyScore)
+LuaC(0x843F20, DeleteCommand)
+LuaC(0x8440A0, GetSpecialFiles)
+LuaC(0x844540, GetSpecialFilePath)
+LuaC(0x844C30, GetSpecialFolder)
+LuaC(0x844F10, RemoveSpecialFile)
+LuaC(0x845540, GetSpecialFileInfo)
+LuaC(0x845DF0, RemoveProfileDirectories)
+LuaC(0x846200, CopyCurrentReplay)
+LuaC(0x846760, SetFrontEndData)
+LuaC(0x846960, GetFrontEndData)
+LuaC(0x846B60, SetOverlayFilter)
+LuaC(0x846F70, SetOverlayFilters)
+LuaC(0x8471C0, GenerateBuildTemplateFromSelection)
+LuaC(0x847270, GetActiveBuildTemplate)
+LuaC(0x847560, SetActiveBuildTemplate)
+LuaC(0x847A20, ClearBuildTemplates)
+LuaC(0x847AD0, RenderOverlayMilitary)
+LuaC(0x847BC0, RenderOverlayIntel)
+LuaC(0x847CB0, RenderOverlayEconomy)
+LuaC(0x847D90, TeamColorMode)
+LuaC(0x847E70, UI_GetUnitById)
+LuaC(0x847FD0, OpenURL)
+LuaC(0x848260, MapBorderAdd)
+LuaC(0x8483D0, MapBorderClear)
+LuaC(0x84DC80, SetCursor)
+LuaC(0x84DDE0, GetCursor)
+LuaC(0x84DEF0, SetUIControlsAlpha)
+LuaC(0x84E000, GetUIControlsAlpha)
+LuaC(0x84E0F0, FlushEvents)
+LuaC(0x850D20, InternalCreateMapPreview)
+LuaC(0x850ED0, CUIMapPreviewSetTexture)
+LuaC(0x851040, CUIMapPreviewSetTextureFromMap)
+LuaC(0x8511B0, CUIMapPreviewClearTexture)
+LuaC(0x857B60, AddCommandFeedbackBlip)
+LuaC(0x866D30, UISelectionByCategory)
+LuaC(0x866F40, UISelectAndZoomTo)
+LuaC(0x8671C0, UIZoomTo)
+LuaC(0x86A8D0, InternalCreateWldUIProvider)
+LuaC(0x86AA50, CLuaWldUIProviderDestroy)
+LuaC(0x86BB30, InternalCreateWorldMesh)
+LuaC(0x86BC90, CUIWorldMeshDestroy)
+LuaC(0x86BDC0, CUIWorldMeshSetMesh)
+LuaC(0x86BF50, CUIWorldMeshSetStance)
+LuaC(0x86C1D0, CUIWorldMeshSetHidden)
+LuaC(0x86C310, CUIWorldMeshIsHidden)
+LuaC(0x86C450, CUIWorldMeshSetAuxiliaryParameter)
+LuaC(0x86C5D0, CUIWorldMeshSetFractionCompleteParameter)
+LuaC(0x86C750, CUIWorldMeshSetFractionHealthParameter)
+LuaC(0x86C8D0, CUIWorldMeshSetLifetimeParameter)
+LuaC(0x86CA50, CUIWorldMeshSetColor)
+LuaC(0x86CBB0, CUIWorldMeshSetScale)
+LuaC(0x86CD40, CUIWorldMeshGetInterpolatedPosition)
+LuaC(0x86CF00, CUIWorldMeshGetInterpolatedSphere)
+LuaC(0x86D0F0, CUIWorldMeshGetInterpolatedAlignedBox)
+LuaC(0x86D320, CUIWorldMeshGetInterpolatedOrientedBox)
+LuaC(0x86D5E0, CUIWorldMeshGetInterpolatedScroll)
+LuaC(0x871690, CUIWorldView__init)
+LuaC(0x871A20, CUIWorldViewCameraReset)
+LuaC(0x871B70, CUIWorldViewGetsGlobalCameraCommands)
+LuaC(0x871CC0, CUIWorldViewGetRightMouseButtonOrder)
+LuaC(0x871EC0, CUIWorldViewHasHighlightCommand)
+LuaC(0x871FA0, CUIWorldViewShowConvertToPatrolCursor)
+LuaC(0x8720E0, CUIWorldViewUnlockInput)
+LuaC(0x872200, CUIWorldViewLockInput)
+LuaC(0x872330, CUIWorldViewIsInputLocked)
+LuaC(0x872470, CUIWorldViewSetHighlightEnabled)
+LuaC(0x8725B0, CUIWorldViewZoomScale)
+LuaC(0x872830, CUIWorldViewGetScreenPos)
+LuaC(0x872C60, UnProject)
+LuaC(0x872E20, CUIWorldViewProject)
+LuaC(0x873170, CUIWorldViewSetCartographic)
+LuaC(0x8732C0, CUIWorldViewIsCartographic)
+LuaC(0x873410, CUIWorldViewEnableResourceRendering)
+LuaC(0x873550, CUIWorldViewIsResourceRenderingEnabled)
+LuaC(0x874FA0, SetMovieVolume)
+LuaC(0x875100, GetMovieVolume)
+LuaC(0x8753A0, GetMovieDuration)
+LuaC(0x876DD0, LaunchReplaySession)
+LuaC(0x87F1E0, _c_CreateDecal)
+LuaC(0x87F360, ScriptedDecalSetTexture)
+LuaC(0x87F4C0, ScriptedDecalSetScale)
+LuaC(0x87F660, ScriptedDecalSetPositionByScreen)
+LuaC(0x87F7E0, ScriptedDecalSetPosition)
+LuaC(0x87F950, ScriptedDecalDestroy)
+LuaC(0x880C40, LoadSavedGame)
+LuaC(0x881AB0, InternalSaveGame)
+LuaC(0x886350, PrefetchSession)
+LuaC(0x88D340, LaunchSinglePlayerSession)
+LuaC(0x88D4C0, GetSessionClients)
+LuaC(0x88D8F0, EjectSessionClient)
+LuaC(0x88DA80, SessionSendChatMessage)
+LuaC(0x88DED0, WorldIsLoading)
+LuaC(0x88DFB0, WorldIsPlaying)
+LuaC(0x88E1E0, GetGameSpeed)
+LuaC(0x88E2E0, SetGameSpeed)
+LuaC(0x897780, SessionRequestPause)
+LuaC(0x897850, SessionResume)
+LuaC(0x897920, SessionIsPaused)
+LuaC(0x897A00, SessionIsGameOver)
+LuaC(0x897AF0, SessionGetCommandSourceNames)
+LuaC(0x897C70, SessionGetLocalCommandSource)
+LuaC(0x897D70, UI_SessionIsReplay)
+LuaC(0x897E60, SessionIsBeingRecorded)
+LuaC(0x897F50, SessionIsMultiplayer)
+LuaC(0x898040, SessionIsObservingAllowed)
+LuaC(0x898130, SessionCanRestart)
+LuaC(0x898220, SessionIsActive)
+LuaC(0x8982F0, SessionGetScenarioInfo)
+LuaC(0x898430, SessionEndGame)
+LuaC(0x8ACDA0, PauseSound)
+LuaC(0x8ACF50, PauseVoice)
+LuaC(0x8AD100, PlaySound)
+LuaC(0x8AD280, SoundIsPrepared)
+LuaC(0x8AD400, StartSound)
+LuaC(0x8AD550, StopSound)
+LuaC(0x8AD6F0, SetVolume)
+LuaC(0x8AD850, GetVolume)
+LuaC(0x8AD970, StopAllSounds)
+LuaC(0x8ADA50, DisableWorldSounds)
+LuaC(0x8ADB30, EnableWorldSounds)
+LuaC(0x8ADC10, PlayTutorialVO)
+LuaC(0x8ADD80, PlayVoice)
+LuaC(0x8ADFF0, UI_AudioSetLanguage)
+LuaC(0x8AE280, UI_HasLocalizedVO)
+LuaC(0x8B9B00, UI_ParseEntityCategory)
+LuaC(0x8B9C90, UI_EntityCategoryContains)
+LuaC(0x8B9EF0, UI_EntityCategoryFilterDown)
+LuaC(0x8BA220, EntityCategoryFilterOut)
+LuaC(0x8BA540, ExecLuaInSim)
+LuaC(0x8BA6F0, SimCallback)
+LuaC(0x8BA9F0, GetIsPaused)
+LuaC(0x8BABA0, GetIsAutoMode)
+LuaC(0x8BAD60, SetAutoMode)
+LuaC(0x8BAF60, GetIsSubmerged)
+LuaC(0x8BB120, GetIsAutoSurfaceMode)
+LuaC(0x8BB2E0, SetAutoSurfaceMode)
+LuaC(0x8BB4E0, GetFireState)
+LuaC(0x8BB6E0, ToggleFireState)
+LuaC(0x8BB990, SetFireState)
+LuaC(0x8BBB90, GetScriptBit)
+LuaC(0x8BBDE0, ToggleScriptBit)
+LuaC(0x8BC080, SetPaused)
+LuaC(0x8BC280, GetAttachedUnitsList)
+LuaC(0x8BC570, ValidateUnitsList)
+LuaC(0x8BC7A0, GetAssistingUnitsList)
+LuaC(0x8BCC30, GetArmyAvatars)
+LuaC(0x8BCE70, GetIdleEngineers)
+LuaC(0x8BD100, GetIdleFactories)
+LuaC(0x8BD390, GetValidAttackingUnits)
+LuaC(0x8BD600, GetSelectedUnits)
+LuaC(0x8BD870, SelectUnits)
+LuaC(0x8BDC30, AddSelectUnits)
+LuaC(0x8BDE40, AddToSessionExtraSelectList)
+LuaC(0x8BDF70, RemoveFromSessionExtraSelectList)
+LuaC(0x8BE0D0, ClearSessionExtraSelectList)
+LuaC(0x8BE1C0, CurrentTime)
+LuaC(0x8BE2A0, GameTime)
+LuaC(0x8BE3A0, GameTick)
+LuaC(0x8BE490, UI_IsAlly)
+LuaC(0x8BE5D0, UI_IsEnemy)
+LuaC(0x8BE710, UI_IsNeutral)
+LuaC(0x8BE850, SyncPlayableRect)
+LuaC(0x8BE980, UI_Random)
+LuaC(0x8C2010, UserUnitCanAttackTarget)
+LuaC(0x8C21D0, UserUnitGetFootPrintSize)
+LuaC(0x8C2340, UserUnitGetUnitId)
+LuaC(0x8C24A0, UserUnitGetEntityId)
+LuaC(0x8C2620, UserUnitGetBlueprint)
+LuaC(0x8C2790, UserUnitHasUnloadCommandQueuedUp)
+LuaC(0x8C29D0, UserUnitProcessInfo)
+LuaC(0x8C2B60, UserUnitIsAutoMode)
+LuaC(0x8C2CA0, UserUnitIsAutoSurfaceMode)
+LuaC(0x8C2DE0, UserUnitIsRepeatQueue)
+LuaC(0x8C2F20, UserUnitIsInCategory)
+LuaC(0x8C30E0, UserUnitGetStat)
+LuaC(0x8C3440, UserUnitIsStunned)
+LuaC(0x8C3580, UserUnitSetCustomName)
+LuaC(0x8C3700, UserUnitGetCustomName)
+LuaC(0x8C3880, UserUnitAddSelectionSet)
+LuaC(0x8C39E0, UserUnitRemoveSelectionSet)
+LuaC(0x8C3B40, UserUnitHasSelectionSet)
+LuaC(0x8C3CD0, UserUnitGetSelectionSets)
+LuaC(0x8C3EA0, UserUnitGetHealth)
+LuaC(0x8C3FE0, UserUnitGetMaxHealth)
+LuaC(0x8C4120, UserUnitGetBuildRate)
+LuaC(0x8C4270, UserUnitIsOverchargePaused)
+LuaC(0x8C43B0, UserUnitIsDead)
+LuaC(0x8C4500, UserUnitIsIdle)
+LuaC(0x8C4660, UserUnitGetFocus)
+LuaC(0x8C47F0, UserUnitGetGuardedEntity)
+LuaC(0x8C4980, UserUnitGetCreator)
+LuaC(0x8C4AF0, UserUnitGetPosition)
+LuaC(0x8C4C50, UserUnitGetArmy)
+LuaC(0x8C4DA0, UserUnitGetFuelRatio)
+LuaC(0x8C4EE0, UserUnitGetShieldRatio)
+LuaC(0x8C5020, UserUnitGetWorkProgress)
+LuaC(0x8C5160, UserUnitGetEconData)
+LuaC(0x8C5400, UserUnitGetCommandQueue)
+LuaC(0x8C5750, UserUnitGetMissileInfo)
+LuaC(0x8C5930, UI_GetBlueprint)
+LuaC(0x8C9490, GetOptions)
+LuaC(0x8C9660, GetAntiAliasingOptions)
+LuaC(0x8C9850, GetPreference)
+LuaC(0x8C9A50, SetPreference)
+LuaC(0x8C9C30, SavePreferences)
+LuaC(0x8CAEA0, DebugFacilitiesEnabled)
+LuaC(0x90A8C0, Lua_LOG)
+LuaC(0x90A8C0, _ALERT)
+LuaC(0x90AF80, import)
+LuaC(0x90AFB0, LuaDumpBinary)
