+++ include/magic_classes.h
